Refactor trego module
